# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Quin Course",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/voucher/check/{code}": {
            "get": {
                "tags": [
                    "Voucher"
                ],
                "summary": "check_voucher ",
                "description": "check_voucher",
                "operationId": "check_voucher",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "code of this course",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Token (Bearer <token>)"
            }
        }
    }
}